scripts/requirements.txt
psycopg2-binary==2.9.7
requests==2.31.0
boto3==1.28.62
